Oracle that gives feedback
Oracle class. Currently has a perfect binary oracle that will always give correct binary feedback (if the action taken is optimal: 1, if the action taken is not optimal: 0), as well as a binary oracle for policy shaping (takes 'likelihood of feedback' as well 'probability that feedback is optimal' parameters). (Yet to be added is a TAMER oracle).

This class assumes the q-table the oracle uses is in pandas DataFrame format (conversion from numpy arrays is fairly straight forward).